feat: Introduce `tsconfig.json` and `.vscode/settings.json`, update exception handling to use `const &`, refine `QCoreApplication::translate` calls, add `static_cast` for type safety, and implement move/copy semantics for GUI components.

+++ core/image/division/image_division.cpp
@@ -35,7 +35,8 @@ bool ImageDivision::saveImpl(bool (*saveFunc)(const QString &, const QImage &, c
 
         for (unsigned int y = 0; y < yMax; y++) {
             for (unsigned int x = 0; x < xMax; x++) {
-                const QRect rect(QPoint(x * cellSize.width(), y * cellSize.height()),
+                const QRect rect(QPoint(static_cast<int>(x * cellSize.width()),
+                                        static_cast<int>(y * cellSize.height())),
                                  cellSize.toSize());
                 const QString filename = saveFilename(dir, x, y);
